refactor(carbon): split speed_consumption_card into part files
Extract `_BarList` + `_SpeedBar` from `speed_consumption_card.dart`
into a sibling `part of` file `_speed_bars.dart`, mirroring the
convention introduced in `gear_inference.dart` (PR #1284). The card
file shrinks from 371 LOC to 162 LOC, comfortably under the 300-LOC
budget tracked in #563.

Layout:
- `speed_consumption_card.dart` (162 LOC) — `library;` + `part`,
  `SpeedConsumptionCard` public widget, `_InsufficientData` empty
  state.
- `_speed_bars.dart` (228 LOC) — `_BarList`, `_SpeedBar`, the
  `_bandLabel` switch.

Public API is unchanged: `SpeedConsumptionCard` still exports from
`speed_consumption_card.dart` with the same constructor signature.
The bar widgets remain library-private via `part of`. The widget
test suite (240 LOC, untouched) passes 5/5 against the new layout.

Refs #563 phase: speed_consumption_card
